requirements.txt
python-telegram-bot==20.0
yt-dlp==2023.03.04
acrcloud
moviepy==1.0.3
requests==2.28.2
unittest2==1.1.0
python-dotenv
pyacrcloud